separate, envs setting, db connection and server + starting in proper way using Promises [ONGOING, WORKING API with 1 end-point]
