+++ src/app.ts
@@ -1,14 +1,30 @@
-import Server from './server';
+import setEnvs from "./setenvs";
+import Server from './Server';
+import DBConnector from "./DBConnector";
 
 /** Code based on https://github.com/CyberZujo/todo-app **/
 
-// Set process.envs, connect DB, set routing, start server
-const starter: Promise<void> = new Server().start()
-  .then((port: unknown) => {
-    console.log(`\n\n👍 👍 👍 👍 👍 👍 👍 👍    SERVER running on port ${port}    👍 👍 👍 👍 👍 👍 👍 👍\n\n`);
-  })
-  .catch((error: string) => {
-    console.error('\n\n⛔ ⛔ ⛔ ⛔ ⛔ ⛔ ⛔ ⛔    Unable to establish server:', error, '    ⛔ ⛔ ⛔ ⛔ ⛔ ⛔ ⛔ ⛔\n\n');
-  });
+// Set process.env
+setEnvs();
+
+// Connect DB, start SERVER
+const starter: Promise<void> =
+// Step 1. DB connection
+  new DBConnector().connect()
+    .then(() => {
+      console.log('\n\n👍 👍 👍 👍 👍 👍 👍 👍    DB connected    👍 👍 👍 👍 👍 👍 👍 👍\n\n');
+// Step 2. SERVER start
+      new Server().start()
+        .then((port: unknown) => {
+          console.log(`\n\n👍 👍 👍 👍 👍 👍 👍 👍    SERVER running on port ${port}    👍 👍 👍 👍 👍 👍 👍 👍\n\n`);
+        })
+        .catch((error: string) => {
+          console.error('\n\n⛔ ⛔ ⛔ ⛔ ⛔ ⛔ ⛔ ⛔    Unable to establish server:', error, '    ⛔ ⛔ ⛔ ⛔ ⛔ ⛔ ⛔ ⛔\n\n');
+        });
+    })
+    .catch((error: Error) => {
+      console.error('\n\n⛔ ⛔ ⛔ ⛔ ⛔ ⛔ ⛔ ⛔    Unable to connect to the DB:', error, '    ⛔ ⛔ ⛔ ⛔ ⛔ ⛔ ⛔ ⛔\n\n');
+      throw error;
+    });
 
 export default starter;
